fix: harden protocol and transport handling
* require explicit HttpMCPServer::begin() so tool registration cannot race in-flight requests; construction no longer starts the listener (breaking change over released 0.3.0)
* validate the JSON-RPC envelope (jsonrpc version, id type, params shape) and require mandatory initialize/tools params per the MCP spec
* enforce an aggregate byte budget on the BLE RX queue (-DMCP_BLE_RX_QUEUE_MAX_BYTES, default 16 KB) and migrate McpBle/BLEMCPServer shared state to std::atomic
* drop the redundant NimBLE service start (2.x starts services when advertising begins)
* reject non-JSON Content-Type on POSTs that carry a body; omit empty schema type from tools/list output
* fix the transport send retry loop that never terminated when max_retries == 255
* bump version to 0.4.0 and point examples at ESP-MCP ^0.4.0

+++ src/mcp_transport.c
@@ -142,7 +142,9 @@ static bool mcp_transport_send_packet(const uint8_t *data, size_t len) {
         return false;
     }
 
-    for (uint8_t attempt = 0; attempt <= s_send_max_retries; attempt++) {
+    /* Use a wider counter than the public uint8_t retry setting. With uint8_t,
+     * max_retries == 255 wrapped attempt back to zero and never terminated. */
+    for (uint16_t attempt = 0; attempt <= (uint16_t)s_send_max_retries; attempt++) {
         int rc = s_send_fn(data, len, s_send_ctx);
         if (rc == 0) {
             return true;
